+++ lib/dune/atom_mapping.ex
@@ -7,23 +7,25 @@ defmodule Dune.AtomMapping do
   @type substitute_string :: String.t()
   @type original_string :: String.t()
   @type sub_mapping :: %{optional(substitute_atom) => original_string}
+  @type extra_info :: %{optional(substitute_atom) => :wrapped}
 
   @typedoc """
   Should be considered opaque
   """
-  @type t :: %__MODULE__{atoms: sub_mapping, modules: sub_mapping}
-  @enforce_keys [:atoms, :modules]
-  defstruct [:atoms, :modules]
+  @type t :: %__MODULE__{atoms: sub_mapping, modules: sub_mapping, extra_info: extra_info}
+  @enforce_keys [:atoms, :modules, :extra_info]
+  defstruct @enforce_keys
 
   @spec new :: t()
   def new do
-    %__MODULE__{atoms: %{}, modules: %{}}
+    %__MODULE__{atoms: %{}, modules: %{}, extra_info: %{}}
   end
 
-  @spec from_atoms([{substitute_atom, original_string}]) :: t
-  def from_atoms(list) when is_list(list) do
+  @spec from_atoms([{substitute_atom, original_string}], [{substitute_atom, :wrapped}]) :: t
+  def from_atoms(list, extra_info) when is_list(list) do
     atoms = build_mapping(list)
-    %__MODULE__{atoms: atoms, modules: %{}}
+    extra_info = Map.new(extra_info)
+    %__MODULE__{atoms: atoms, modules: %{}, extra_info: extra_info}
   end
 
   @spec add_modules(t, [{substitute_atom, original_string}]) :: t
@@ -42,6 +44,7 @@ defmodule Dune.AtomMapping do
   def to_string(mapping = %__MODULE__{}, atom) when is_atom(atom) do
     case lookup_original_string(mapping, atom) do
       {:atom, string} -> string
+      {:wrapped_atom, string} -> string
       {:module, string} -> "Elixir.#{string}"
       :error -> Atom.to_string(atom)
     end
@@ -49,15 +52,15 @@ defmodule Dune.AtomMapping do
 
   @spec inspect(t, atom) :: String.t()
   def inspect(mapping = %__MODULE__{}, atom) when is_atom(atom) do
-    # TODO might need to check escaped atoms
     case lookup_original_string(mapping, atom) do
       {:atom, string} -> ":#{string}"
+      {:wrapped_atom, string} -> ~s(:"#{string}")
       {:module, string} -> string
       :error -> inspect(atom)
     end
   end
 
-  @spec lookup_original_string(t, atom) :: {:atom, String.t()} | {:module, String.t()} | :error
+  @spec lookup_original_string(t, atom) :: {:atom | :wrapped_atom | :module, String.t()} | :error
   def lookup_original_string(mapping = %__MODULE__{}, atom) when is_atom(atom) do
     case mapping.modules do
       %{^atom => string} ->
@@ -65,8 +68,14 @@ defmodule Dune.AtomMapping do
 
       _ ->
         case mapping.atoms do
-          %{^atom => string} -> {:atom, string}
-          _ -> :error
+          %{^atom => string} ->
+            case mapping.extra_info do
+              %{^atom => :wrapped} -> {:wrapped_atom, string}
+              _ -> {:atom, string}
+            end
+
+          _ ->
+            :error
         end
     end
   end