+++ test/dune_string_test.exs
@@ -230,6 +230,20 @@ defmodule DuneStringTest do
       assert %Success{value: Dune_Module_1__, inspected: ~s'Elixir.Elixir'} = ~E':"Elixir.Elixir"'
     end
 
+    test "atoms (wrapped with quotes)" do
+      assert %Success{value: :__Dune_atom_1__, inspected: ~s':" "'} = ~E':" "'
+      assert %Success{value: :__Dune_atom_1__, inspected: ~s':"foo/bar"'} = ~E':"foo/bar"'
+
+      assert %Success{value: " ", inspected: ~s'" "'} = ~E'to_string :" "'
+      assert %Success{value: "foo/bar", inspected: ~s'"foo/bar"'} = ~E'to_string :"foo/bar"'
+
+      # FIXME: fix test once Elixir tokenizer is fixed
+      # assert %Success{
+      #          value: ["TODO"],
+      #          inspected: ~s([" ": {:"\t", :" A"}, "ab cd": %{abc: 6, "foo+91": :'15'}])
+      #        } = ~E([" ": {:"\t", :" A"}, "ab cd": %{abc: 6, "foo+91": :'15'}])
+    end
+
     test "function and atom parameters" do
       assert ":digits" = ~E':digits'.value |> inspect()
       assert ":turkic" = ~E':turkic'.value |> inspect()