+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   `static_atoms_encoder`
 - Handle Elixir 1.12 bug on single atom ASTs
 - Handle atoms prefixed with `Elixir.` properly
+- Fix inspect for quoted atoms
 
 ## v0.1.0 (2021-09-19)